Move the sub resolve logic to common place
