+++ components/org.wso2.carbon.identity.oauth/src/main/java/org/wso2/carbon/identity/oauth2/token/AccessTokenIssuer.java
@@ -26,8 +26,16 @@ import org.apache.commons.logging.LogFactory;
 import org.apache.oltu.oauth2.common.error.OAuthError;
 import org.apache.oltu.oauth2.common.message.types.GrantType;
 import org.owasp.encoder.Encode;
+import org.wso2.carbon.identity.application.authentication.framework.exception.UserIdNotFoundException;
+import org.wso2.carbon.identity.application.authentication.framework.model.AuthenticatedUser;
+import org.wso2.carbon.identity.application.common.IdentityApplicationManagementException;
+import org.wso2.carbon.identity.application.common.model.ClaimConfig;
+import org.wso2.carbon.identity.application.common.model.ClaimMapping;
+import org.wso2.carbon.identity.application.common.model.ServiceProvider;
+import org.wso2.carbon.identity.application.common.model.ServiceProviderProperty;
 import org.wso2.carbon.identity.base.IdentityConstants;
 import org.wso2.carbon.identity.base.IdentityException;
+import org.wso2.carbon.identity.core.util.IdentityTenantUtil;
 import org.wso2.carbon.identity.core.util.IdentityUtil;
 import org.wso2.carbon.identity.oauth.cache.AppInfoCache;
 import org.wso2.carbon.identity.oauth.cache.AuthorizationGrantCache;
@@ -54,6 +62,8 @@ import org.wso2.carbon.identity.oauth2.util.OAuth2Util;
 import org.wso2.carbon.identity.oauth2.validators.JDBCPermissionBasedInternalScopeValidator;
 import org.wso2.carbon.identity.oauth2.validators.RoleBasedInternalScopeValidator;
 import org.wso2.carbon.identity.openidconnect.IDTokenBuilder;
+import org.wso2.carbon.user.core.common.AbstractUserStoreManager;
+import org.wso2.carbon.user.core.util.UserCoreUtil;
 import org.wso2.carbon.utils.CarbonUtils;
 
 import java.util.ArrayList;
@@ -63,6 +73,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 
+import static org.apache.commons.lang.StringUtils.isNotBlank;
 import static org.wso2.carbon.identity.oauth.common.OAuthConstants.GrantTypes.REFRESH_TOKEN;
 import static org.wso2.carbon.identity.oauth.common.OAuthConstants.OauthAppStates.APP_STATE_ACTIVE;
 import static org.wso2.carbon.identity.oauth2.Oauth2ScopeConstants.CONSOLE_SCOPE_PREFIX;
@@ -348,6 +359,13 @@ public class AccessTokenIssuer {
             // set the token request context to be used by downstream handlers. This is introduced as a fix for
             // IDENTITY-4111.
             OAuth2Util.setTokenRequestContext(tokReqMsgCtx);
+
+            AuthenticatedUser authorizedUser = tokReqMsgCtx.getAuthorizedUser();
+            if (authorizedUser.getAuthenticatedSubjectIdentifier() == null) {
+                authorizedUser.setAuthenticatedSubjectIdentifier(
+                        getSubjectClaim(getServiceProvider(tokReqMsgCtx.getOauth2AccessTokenReqDTO()), authorizedUser));
+            }
+
             tokenRespDTO = authzGrantHandler.issue(tokReqMsgCtx);
             if (tokenRespDTO.isError()) {
                 setResponseHeaders(tokReqMsgCtx, tokenRespDTO);
@@ -404,6 +422,164 @@ public class AccessTokenIssuer {
         }
 
         return tokenRespDTO;
+    }
+
+    private ServiceProvider getServiceProvider(OAuth2AccessTokenReqDTO tokenReq) throws IdentityOAuth2Exception {
+        ServiceProvider serviceProvider;
+        try {
+            serviceProvider = OAuth2ServiceComponentHolder.getApplicationMgtService().getServiceProviderByClientId(
+                    tokenReq.getClientId(), OAuthConstants.Scope.OAUTH2, tokenReq.getTenantDomain());
+        } catch (IdentityApplicationManagementException e) {
+            throw new IdentityOAuth2Exception("Error occurred while retrieving OAuth2 application data for client id " +
+                    tokenReq.getClientId(), e);
+        }
+        if (serviceProvider == null) {
+            if (log.isDebugEnabled()) {
+                log.debug("Could not find an application for client id: " + tokenReq.getClientId()
+                        + ", scope: " + OAuthConstants.Scope.OAUTH2 + ", tenant: " + tokenReq.getTenantDomain());
+            }
+            throw new IdentityOAuth2Exception("Service Provider not found");
+        }
+        if (log.isDebugEnabled()) {
+            log.debug("Retrieved service provider: " + serviceProvider.getApplicationName() + " for client: " +
+                    tokenReq.getClientId() + ", scope: " + OAuthConstants.Scope.OAUTH2 + ", tenant: " +
+                    tokenReq.getTenantDomain());
+        }
+
+        return serviceProvider;
+    }
+
+    private String getSubjectClaim(ServiceProvider serviceProvider,
+                                   AuthenticatedUser authenticatedUser) throws IdentityOAuth2Exception {
+
+        String userTenantDomain = authenticatedUser.getTenantDomain();
+        String subject;
+        String userStoreDomain = authenticatedUser.getUserStoreDomain();
+        String subjectClaimUri = getSubjectClaimUriInLocalDialect(serviceProvider);
+        if (StringUtils.isNotBlank(subjectClaimUri)) {
+            try {
+                subject = getSubjectClaimFromUserStore(subjectClaimUri, authenticatedUser);
+                if (StringUtils.isBlank(subject)) {
+                    // Set username as the subject claim since we have no other option
+                    subject = getDefaultSubject(serviceProvider, authenticatedUser);
+                    log.warn("Cannot find subject claim: " + subjectClaimUri + " for user:"
+                            + authenticatedUser.getLoggableUserId()
+                            + ". Defaulting to username: " + subject + " as the subject identifier.");
+                }
+                // Get the subject claim in the correct format (ie. tenantDomain or userStoreDomain appended)
+                subject = getFormattedSubjectClaim(serviceProvider, subject, userStoreDomain, userTenantDomain);
+            } catch (IdentityException e) {
+                String error = "Error occurred while getting user claim for user: "
+                        + authenticatedUser.getLoggableUserId() + ", claim" +
+                        ": " +
+                        subjectClaimUri;
+                throw new IdentityOAuth2Exception(error, e);
+            } catch (org.wso2.carbon.user.core.UserStoreException e) {
+                String error = "Error occurred while getting subject claim: " + subjectClaimUri + " for user: "
+                        + authenticatedUser.getLoggableUserId();
+                throw new IdentityOAuth2Exception(error, e);
+            }
+        } else {
+            try {
+                subject = getDefaultSubject(serviceProvider, authenticatedUser);
+                subject = getFormattedSubjectClaim(serviceProvider, subject, userStoreDomain, userTenantDomain);
+            } catch (UserIdNotFoundException e) {
+                throw new IdentityOAuth2Exception("User id not found for user: "
+                        + authenticatedUser.getLoggableUserId(), e);
+            }
+            if (log.isDebugEnabled()) {
+                log.debug("No subject claim defined for service provider: " + serviceProvider.getApplicationName()
+                        + ". Using username as the subject claim.");
+            }
+
+        }
+        return subject;
+    }
+
+    private String getDefaultSubject(ServiceProvider serviceProvider, AuthenticatedUser authenticatedUser)
+            throws UserIdNotFoundException {
+        String subject;
+        boolean useUserIdForSubject = false;
+        ServiceProviderProperty[] spProperties = serviceProvider.getSpProperties();
+        if (spProperties != null) {
+            for (ServiceProviderProperty prop : spProperties) {
+                if ("useUserIdForSubject".equals(prop.getName())) {
+                    useUserIdForSubject = Boolean.parseBoolean(prop.getValue());
+                    break;
+                }
+            }
+        }
+        if (useUserIdForSubject) {
+            subject = authenticatedUser.getUserId();
+        } else {
+            subject = authenticatedUser.getUserName();
+        }
+        return subject;
+    }
+
+    private String getFormattedSubjectClaim(ServiceProvider serviceProvider, String subjectClaimValue,
+                                            String userStoreDomain, String tenantDomain) {
+
+        boolean appendUserStoreDomainToSubjectClaim = serviceProvider.getLocalAndOutBoundAuthenticationConfig()
+                .isUseUserstoreDomainInLocalSubjectIdentifier();
+
+        boolean appendTenantDomainToSubjectClaim = serviceProvider.getLocalAndOutBoundAuthenticationConfig()
+                .isUseTenantDomainInLocalSubjectIdentifier();
+
+        if (appendTenantDomainToSubjectClaim) {
+            subjectClaimValue = UserCoreUtil.addTenantDomainToEntry(subjectClaimValue, tenantDomain);
+        }
+        if (appendUserStoreDomainToSubjectClaim) {
+            subjectClaimValue = IdentityUtil.addDomainToName(subjectClaimValue, userStoreDomain);
+        }
+
+        return subjectClaimValue;
+    }
+
+    private String getSubjectClaimFromUserStore(String subjectClaimUri, AuthenticatedUser authenticatedUser)
+            throws org.wso2.carbon.user.core.UserStoreException, IdentityException {
+
+        AbstractUserStoreManager userStoreManager = (AbstractUserStoreManager) IdentityTenantUtil
+                .getRealm(authenticatedUser.getTenantDomain(), authenticatedUser.toFullQualifiedUsername())
+                .getUserStoreManager();
+
+        return userStoreManager
+                .getUserClaimValueWithID(authenticatedUser.getUserId(), subjectClaimUri, null);
+    }
+
+    private String getSubjectClaimUriInLocalDialect(ServiceProvider serviceProvider) {
+
+        String subjectClaimUri = serviceProvider.getLocalAndOutBoundAuthenticationConfig().getSubjectClaimUri();
+        if (log.isDebugEnabled()) {
+            if (isNotBlank(subjectClaimUri)) {
+                log.debug(subjectClaimUri + " is defined as subject claim for service provider: " +
+                        serviceProvider.getApplicationName());
+            } else {
+                log.debug("No subject claim defined for service provider: " + serviceProvider.getApplicationName());
+            }
+        }
+        // Get the local subject claim URI, if subject claim was a SP mapped one
+        return getSubjectClaimUriInLocalDialect(serviceProvider, subjectClaimUri);
+    }
+
+    private String getSubjectClaimUriInLocalDialect(ServiceProvider serviceProvider, String subjectClaimUri) {
+
+        if (isNotBlank(subjectClaimUri)) {
+            ClaimConfig claimConfig = serviceProvider.getClaimConfig();
+            if (claimConfig != null) {
+                boolean isLocalClaimDialect = claimConfig.isLocalClaimDialect();
+                ClaimMapping[] claimMappings = claimConfig.getClaimMappings();
+                if (!isLocalClaimDialect && ArrayUtils.isNotEmpty(claimMappings)) {
+                    for (ClaimMapping claimMapping : claimMappings) {
+                        if (StringUtils.equals(claimMapping.getRemoteClaim().getClaimUri(), subjectClaimUri)) {
+                            return claimMapping.getLocalClaim().getClaimUri();
+                        }
+                    }
+                }
+            }
+        }
+        // This means the original subjectClaimUri passed was the subject claim URI.
+        return subjectClaimUri;
     }
 
     private void addAuthorizedInternalScopes(OAuthTokenReqMessageContext tokReqMsgCtx,
